second Commit
Forgot to work on Git repo rather than local copy. Updated a lot...

Able to add files to the CWD (still cant download because the
LinkButtons are sending the wrong data back)

Forgot to open my git version of the project instead of my local
version. Hence the large amount of new stuff.